requirements.txt
boto==2.39.0
credstash==1.10.0
requests==2.9.1
